break down parsing
