+++ aws/logs_monitoring/steps/parsing.py
@@ -3,70 +3,32 @@
 # This product includes software developed at Datadog (https://www.datadoghq.com/).
 # Copyright 2021 Datadog, Inc.
 
-import base64
-import gzip
 import json
 import os
-import boto3
-import botocore
 import itertools
-import re
-import urllib
 import logging
-from io import BytesIO, BufferedReader
 from datadog_lambda.metric import lambda_stats
-from customized_log_group import (
-    get_lambda_function_name_from_logstream_name,
-    is_lambda_customized_log_group,
-)
-from caching.step_functions_cache import StepFunctionsTagsCache
-from caching.cloudwatch_log_group_cache import CloudwatchLogGroupTagsCache
 from telemetry import (
     DD_FORWARDER_TELEMETRY_NAMESPACE_PREFIX,
     get_forwarder_telemetry_tags,
     set_forwarder_telemetry_tags,
 )
+from steps.handlers.awslogs_handler import awslogs_handler
+from steps.handlers.s3_handler import s3_handler
+from steps.common import (
+    merge_dicts,
+    get_service_from_tags_and_remove_duplicates,
+)
 from settings import (
     DD_TAGS,
-    DD_MULTILINE_LOG_REGEX_PATTERN,
     DD_SOURCE,
     DD_CUSTOM_TAGS,
     DD_SERVICE,
-    DD_HOST,
     DD_FORWARDER_VERSION,
-    DD_USE_VPC,
 )
-
-GOV, CN = "gov", "cn"
 
 logger = logging.getLogger()
-
-if DD_MULTILINE_LOG_REGEX_PATTERN:
-    try:
-        multiline_regex = re.compile(
-            "[\n\r\f]+(?={})".format(DD_MULTILINE_LOG_REGEX_PATTERN)
-        )
-    except Exception:
-        raise Exception(
-            "could not compile multiline regex with pattern: {}".format(
-                DD_MULTILINE_LOG_REGEX_PATTERN
-            )
-        )
-    multiline_regex_start_pattern = re.compile(
-        "^{}".format(DD_MULTILINE_LOG_REGEX_PATTERN)
-    )
-
-rds_regex = re.compile("/aws/rds/(instance|cluster)/(?P<host>[^/]+)/(?P<name>[^/]+)")
-
-cloudtrail_regex = re.compile(
-    "\d+_CloudTrail(|-Digest)_\w{2}(|-gov|-cn)-\w{4,9}-\d_(|.+)\d{8}T\d{4,6}Z(|.+).json.gz$",
-    re.I,
-)
-
-# Store the cache in the global scope so that it will be reused as long as
-# the log forwarder Lambda container is running
-account_cw_logs_tags_cache = CloudwatchLogGroupTagsCache()
-account_step_functions_tags_cache = StepFunctionsTagsCache()
+logger.setLevel(logging.getLevelName(os.environ.get("DD_LOG_LEVEL", "INFO").upper()))
 
 
 def parse(event, context):
@@ -114,7 +76,6 @@ def generate_metadata(context):
         "forwarder_memorysize": context.memory_limit_in_mb,
         "forwarder_version": DD_FORWARDER_VERSION,
     }
-
     metadata[DD_CUSTOM_TAGS] = ",".join(
         filter(
             None,
@@ -155,490 +116,6 @@ def parse_event_type(event):
     elif "detail" in event:
         return "events"
     raise Exception("Event type not supported (see #Event supported section)")
-
-
-# Handle S3 events
-def s3_handler(event, context, metadata):
-    # Need to use path style to access s3 via VPC Endpoints
-    # https://github.com/gford1000-aws/lambda_s3_access_using_vpc_endpoint#boto3-specific-notes
-    if DD_USE_VPC:
-        s3 = boto3.client(
-            "s3",
-            os.environ["AWS_REGION"],
-            config=botocore.config.Config(s3={"addressing_style": "path"}),
-        )
-    else:
-        s3 = boto3.client("s3")
-    # if this is a S3 event carried in a SNS message, extract it and override the event
-    if "Sns" in event["Records"][0]:
-        event = json.loads(event["Records"][0]["Sns"]["Message"])
-
-    # Get the object from the event and show its content type
-    bucket = event["Records"][0]["s3"]["bucket"]["name"]
-    key = urllib.parse.unquote_plus(event["Records"][0]["s3"]["object"]["key"])
-
-    source = parse_event_source(event, key)
-    if "transit-gateway" in bucket:
-        source = "transitgateway"
-    metadata[DD_SOURCE] = source
-
-    metadata[DD_SERVICE] = get_service_from_tags_and_remove_duplicates(metadata)
-
-    ##Get the ARN of the service and set it as the hostname
-    hostname = parse_service_arn(source, key, bucket, context)
-    if hostname:
-        metadata[DD_HOST] = hostname
-
-    # Extract the S3 object
-    response = s3.get_object(Bucket=bucket, Key=key)
-    body = response["Body"]
-    data = body.read()
-
-    yield from get_structured_lines_for_s3_handler(data, bucket, key, source)
-
-
-def get_structured_lines_for_s3_handler(data, bucket, key, source):
-    # Decompress data that has a .gz extension or magic header http://www.onicos.com/staff/iz/formats/gzip.html
-    if key[-3:] == ".gz" or data[:2] == b"\x1f\x8b":
-        with gzip.GzipFile(fileobj=BytesIO(data)) as decompress_stream:
-            # Reading line by line avoid a bug where gzip would take a very long time (>5min) for
-            # file around 60MB gzipped
-            data = b"".join(BufferedReader(decompress_stream))
-
-    is_cloudtrail_bucket = False
-    if is_cloudtrail(str(key)):
-        try:
-            cloud_trail = json.loads(data)
-            if cloud_trail.get("Records") is not None:
-                # only parse as a cloudtrail bucket if we have a Records field to parse
-                is_cloudtrail_bucket = True
-                for event in cloud_trail["Records"]:
-                    # Create structured object and send it
-                    structured_line = merge_dicts(
-                        event, {"aws": {"s3": {"bucket": bucket, "key": key}}}
-                    )
-                    yield structured_line
-        except Exception as e:
-            logger.debug("Unable to parse cloudtrail log: %s" % e)
-
-    if not is_cloudtrail_bucket:
-        # Check if using multiline log regex pattern
-        # and determine whether line or pattern separated logs
-        data = data.decode("utf-8", errors="ignore")
-        if DD_MULTILINE_LOG_REGEX_PATTERN and multiline_regex_start_pattern.match(data):
-            split_data = multiline_regex.split(data)
-        else:
-            if DD_MULTILINE_LOG_REGEX_PATTERN:
-                logger.debug(
-                    "DD_MULTILINE_LOG_REGEX_PATTERN %s did not match start of file, splitting by line",
-                    DD_MULTILINE_LOG_REGEX_PATTERN,
-                )
-            if source == "waf":
-                # WAF logs are \n separated
-                split_data = [d for d in data.split("\n") if d != ""]
-            else:
-                split_data = data.splitlines()
-
-        # Send lines to Datadog
-        for line in split_data:
-            # Create structured object and send it
-            structured_line = {
-                "aws": {"s3": {"bucket": bucket, "key": key}},
-                "message": line,
-            }
-            yield structured_line
-
-
-def get_service_from_tags_and_remove_duplicates(metadata):
-    service = ""
-    tagsplit = metadata[DD_CUSTOM_TAGS].split(",")
-    for i, tag in enumerate(tagsplit):
-        if tag.startswith("service:"):
-            if service:
-                # remove duplicate entry from the tags
-                del tagsplit[i]
-            else:
-                service = tag[8:]
-
-    metadata[DD_CUSTOM_TAGS] = ",".join(tagsplit)
-
-    # Default service to source value
-    return service if service else metadata[DD_SOURCE]
-
-
-def parse_event_source(event, key):
-    """Parse out the source that will be assigned to the log in Datadog
-    Args:
-        event (dict): The AWS-formatted log event that the forwarder was triggered with
-        key (string): The S3 object key if the event is from S3 or the CW Log Group if the event is from CW Logs
-    """
-    lowercase_key = str(key).lower()
-
-    # Determines if the key matches any known sources for Cloudwatch logs
-    if "awslogs" in event:
-        return find_cloudwatch_source(lowercase_key)
-
-    # Determines if the key matches any known sources for S3 logs
-    if "Records" in event and len(event["Records"]) > 0:
-        if "s3" in event["Records"][0]:
-            if is_cloudtrail(str(key)):
-                return "cloudtrail"
-
-            return find_s3_source(lowercase_key)
-
-    return "aws"
-
-
-def find_cloudwatch_source(log_group):
-    # e.g. /aws/rds/instance/my-mariadb/error
-    if log_group.startswith("/aws/rds"):
-        for engine in ["mariadb", "mysql", "postgresql"]:
-            if engine in log_group:
-                return engine
-        return "rds"
-
-    if log_group.startswith(
-        (
-            # default location for rest api execution logs
-            "api-gateway",  # e.g. Api-Gateway-Execution-Logs_xxxxxx/dev
-            # default location set by serverless framework for rest api access logs
-            "/aws/api-gateway",  # e.g. /aws/api-gateway/my-project
-            # default location set by serverless framework for http api logs
-            "/aws/http-api",  # e.g. /aws/http-api/my-project
-        )
-    ):
-        return "apigateway"
-
-    if log_group.startswith("/aws/vendedlogs/states"):
-        return "stepfunction"
-
-    # e.g. dms-tasks-test-instance
-    if log_group.startswith("dms-tasks"):
-        return "dms"
-
-    # e.g. sns/us-east-1/123456779121/SnsTopicX
-    if log_group.startswith("sns/"):
-        return "sns"
-
-    # e.g. /aws/fsx/windows/xxx
-    if log_group.startswith("/aws/fsx/windows"):
-        return "aws.fsx"
-
-    if log_group.startswith("/aws/appsync/"):
-        return "appsync"
-
-    for source in [
-        "/aws/lambda",  # e.g. /aws/lambda/helloDatadog
-        "/aws/codebuild",  # e.g. /aws/codebuild/my-project
-        "/aws/kinesis",  # e.g. /aws/kinesisfirehose/dev
-        "/aws/docdb",  # e.g. /aws/docdb/yourClusterName/profile
-        "/aws/eks",  # e.g. /aws/eks/yourClusterName/profile
-    ]:
-        if log_group.startswith(source):
-            return source.replace("/aws/", "")
-
-    # the below substrings must be in your log group to be detected
-    for source in [
-        "network-firewall",
-        "route53",
-        "vpc",
-        "fargate",
-        "cloudtrail",
-        "msk",
-        "elasticsearch",
-        "transitgateway",
-        "verified-access",
-        "bedrock",
-    ]:
-        if source in log_group:
-            return source
-
-    return "cloudwatch"
-
-
-def is_cloudtrail(key):
-    match = cloudtrail_regex.search(key)
-    return bool(match)
-
-
-def find_s3_source(key):
-    # e.g. AWSLogs/123456779121/elasticloadbalancing/us-east-1/2020/10/02/123456779121_elasticloadbalancing_us-east-1_app.alb.xxxxx.xx.xxx.xxx_x.log.gz
-    if "elasticloadbalancing" in key:
-        return "elb"
-
-    # e.g. AWSLogs/123456779121/vpcflowlogs/us-east-1/2020/10/02/123456779121_vpcflowlogs_us-east-1_fl-xxxxx.log.gz
-    if "vpcflowlogs" in key:
-        return "vpc"
-
-    # e.g. AWSLogs/123456779121/vpcdnsquerylogs/vpc-********/2021/05/11/vpc-********_vpcdnsquerylogs_********_20210511T0910Z_71584702.log.gz
-    if "vpcdnsquerylogs" in key:
-        return "route53"
-
-    # e.g. 2020/10/02/21/aws-waf-logs-testing-1-2020-10-02-21-25-30-x123x-x456x or AWSLogs/123456779121/WAFLogs/us-east-1/xxxxxx-waf/2022/10/11/14/10/123456779121_waflogs_us-east-1_xxxxx-waf_20221011T1410Z_12756524.log.gz
-    if "aws-waf-logs" in key or "waflogs" in key:
-        return "waf"
-
-    # e.g. AWSLogs/123456779121/redshift/us-east-1/2020/10/21/123456779121_redshift_us-east-1_mycluster_userlog_2020-10-21T18:01.gz
-    if "_redshift_" in key:
-        return "redshift"
-
-    # this substring must be in your target prefix to be detected
-    if "amazon_documentdb" in key:
-        return "docdb"
-
-    # e.g. carbon-black-cloud-forwarder/alerts/org_key=*****/year=2021/month=7/day=19/hour=18/minute=15/second=41/8436e850-7e78-40e4-b3cd-6ebbc854d0a2.jsonl.gz
-    if "carbon-black" in key:
-        return "carbonblack"
-
-    # the below substrings must be in your target prefix to be detected
-    for source in [
-        "amazon_codebuild",
-        "amazon_kinesis",
-        "amazon_dms",
-        "amazon_msk",
-        "network-firewall",
-        "cloudfront",
-        "verified-access",
-        "bedrock",
-    ]:
-        if source in key:
-            return source.replace("amazon_", "")
-
-    return "s3"
-
-
-def get_partition_from_region(region):
-    partition = "aws"
-    if region:
-        if GOV in region:
-            partition = "aws-us-gov"
-        elif CN in region:
-            partition = "aws-cn"
-    return partition
-
-
-def parse_service_arn(source, key, bucket, context):
-    if source == "elb":
-        # For ELB logs we parse the filename to extract parameters in order to rebuild the ARN
-        # 1. We extract the region from the filename
-        # 2. We extract the loadbalancer name and replace the "." by "/" to match the ARN format
-        # 3. We extract the id of the loadbalancer
-        # 4. We build the arn
-        idsplit = key.split("/")
-        if not idsplit:
-            logger.debug("Invalid service ARN, unable to parse ELB ARN")
-            return
-        # If there is a prefix on the S3 bucket, remove the prefix before splitting the key
-        if idsplit[0] != "AWSLogs":
-            try:
-                idsplit = idsplit[idsplit.index("AWSLogs") :]
-                keysplit = "/".join(idsplit).split("_")
-            except ValueError:
-                logger.debug("Invalid S3 key, doesn't contain AWSLogs")
-                return
-        # If no prefix, split the key
-        else:
-            keysplit = key.split("_")
-        if len(keysplit) > 3:
-            region = keysplit[2].lower()
-            name = keysplit[3]
-            elbname = name.replace(".", "/")
-            if len(idsplit) > 1:
-                idvalue = idsplit[1]
-                partition = get_partition_from_region(region)
-                return "arn:{}:elasticloadbalancing:{}:{}:loadbalancer/{}".format(
-                    partition, region, idvalue, elbname
-                )
-    if source == "s3":
-        # For S3 access logs we use the bucket name to rebuild the arn
-        if bucket:
-            return "arn:aws:s3:::{}".format(bucket)
-    if source == "cloudfront":
-        # For Cloudfront logs we need to get the account and distribution id from the lambda arn and the filename
-        # 1. We extract the cloudfront id  from the filename
-        # 2. We extract the AWS account id from the lambda arn
-        # 3. We build the arn
-        namesplit = key.split("/")
-        if len(namesplit) > 0:
-            filename = namesplit[len(namesplit) - 1]
-            # (distribution-ID.YYYY-MM-DD-HH.unique-ID.gz)
-            filenamesplit = filename.split(".")
-            if len(filenamesplit) > 3:
-                distributionID = filenamesplit[len(filenamesplit) - 4].lower()
-                arn = context.invoked_function_arn
-                arnsplit = arn.split(":")
-                if len(arnsplit) == 7:
-                    awsaccountID = arnsplit[4].lower()
-                    return "arn:aws:cloudfront::{}:distribution/{}".format(
-                        awsaccountID, distributionID
-                    )
-    if source == "redshift":
-        # For redshift logs we leverage the filename to extract the relevant information
-        # 1. We extract the region from the filename
-        # 2. We extract the account-id from the filename
-        # 3. We extract the name of the cluster
-        # 4. We build the arn: arn:aws:redshift:region:account-id:cluster:cluster-name
-        namesplit = key.split("/")
-        if len(namesplit) == 8:
-            region = namesplit[3].lower()
-            accountID = namesplit[1].lower()
-            filename = namesplit[7]
-            filesplit = filename.split("_")
-            if len(filesplit) == 6:
-                clustername = filesplit[3]
-                return "arn:{}:redshift:{}:{}:cluster:{}:".format(
-                    get_partition_from_region(region), region, accountID, clustername
-                )
-    return
-
-
-# Handle CloudWatch logs
-def awslogs_handler(event, context, metadata):
-    # Get logs
-    with gzip.GzipFile(
-        fileobj=BytesIO(base64.b64decode(event["awslogs"]["data"]))
-    ) as decompress_stream:
-        # Reading line by line avoid a bug where gzip would take a very long
-        # time (>5min) for file around 60MB gzipped
-        data = b"".join(BufferedReader(decompress_stream))
-    logs = json.loads(data)
-
-    # Set the source on the logs
-    source = logs.get("logGroup", "cloudwatch")
-
-    # Use the logStream to identify if this is a CloudTrail, TransitGateway, or Bedrock event
-    # i.e. 123456779121_CloudTrail_us-east-1
-    if "_CloudTrail_" in logs["logStream"]:
-        source = "cloudtrail"
-    if "tgw-attach" in logs["logStream"]:
-        source = "transitgateway"
-    if logs["logStream"] == "aws/bedrock/modelinvocations":
-        source = "bedrock"
-    metadata[DD_SOURCE] = parse_event_source(event, source)
-
-    # Special handling for customized log group of Lambda functions
-    # Multiple Lambda functions can share one single customized log group
-    # Need to parse logStream name to determine whether it is a Lambda function
-    if is_lambda_customized_log_group(logs["logStream"]):
-        metadata[DD_SOURCE] = "lambda"
-
-    # Build aws attributes
-    aws_attributes = {
-        "aws": {
-            "awslogs": {
-                "logGroup": logs["logGroup"],
-                "logStream": logs["logStream"],
-                "owner": logs["owner"],
-            }
-        }
-    }
-
-    formatted_tags = account_cw_logs_tags_cache.get(logs["logGroup"])
-    if len(formatted_tags) > 0:
-        metadata[DD_CUSTOM_TAGS] = (
-            ",".join(formatted_tags)
-            if not metadata[DD_CUSTOM_TAGS]
-            else metadata[DD_CUSTOM_TAGS] + "," + ",".join(formatted_tags)
-        )
-
-    # Set service from custom tags, which may include the tags set on the log group
-    # Returns DD_SOURCE by default
-    metadata[DD_SERVICE] = get_service_from_tags_and_remove_duplicates(metadata)
-
-    # Set host as log group where cloudwatch is source
-    if metadata[DD_SOURCE] == "cloudwatch" or metadata.get(DD_HOST, None) == None:
-        metadata[DD_HOST] = aws_attributes["aws"]["awslogs"]["logGroup"]
-
-    if metadata[DD_SOURCE] == "appsync":
-        metadata[DD_HOST] = aws_attributes["aws"]["awslogs"]["logGroup"].split("/")[-1]
-
-    if metadata[DD_SOURCE] == "verified-access":
-        try:
-            message = json.loads(logs["logEvents"][0]["message"])
-            metadata[DD_HOST] = message["http_request"]["url"]["hostname"]
-        except Exception as e:
-            logger.debug("Unable to set verified-access log host: %s" % e)
-
-    if metadata[DD_SOURCE] == "stepfunction" and logs["logStream"].startswith(
-        "states/"
-    ):
-        state_machine_arn = ""
-        try:
-            state_machine_arn = get_state_machine_arn(
-                json.loads(logs["logEvents"][0]["message"])
-            )
-            if state_machine_arn:  # not empty
-                metadata[DD_HOST] = state_machine_arn
-        except Exception as e:
-            logger.debug(
-                "Unable to set stepfunction host or get state_machine_arn: %s" % e
-            )
-
-        formatted_stepfunctions_tags = account_step_functions_tags_cache.get(
-            state_machine_arn
-        )
-        if len(formatted_stepfunctions_tags) > 0:
-            metadata[DD_CUSTOM_TAGS] = (
-                ",".join(formatted_stepfunctions_tags)
-                if not metadata[DD_CUSTOM_TAGS]
-                else metadata[DD_CUSTOM_TAGS]
-                + ","
-                + ",".join(formatted_stepfunctions_tags)
-            )
-
-    # When parsing rds logs, use the cloudwatch log group name to derive the
-    # rds instance name, and add the log name of the stream ingested
-    if metadata[DD_SOURCE] in ["rds", "mariadb", "mysql", "postgresql"]:
-        match = rds_regex.match(logs["logGroup"])
-        if match is not None:
-            metadata[DD_HOST] = match.group("host")
-            metadata[DD_CUSTOM_TAGS] = (
-                metadata[DD_CUSTOM_TAGS] + ",logname:" + match.group("name")
-            )
-
-    # For Lambda logs we want to extract the function name,
-    # then rebuild the arn of the monitored lambda using that name.
-    if metadata[DD_SOURCE] == "lambda":
-        process_lambda_logs(logs, aws_attributes, context, metadata)
-
-    # The EKS log group contains various sources from the K8S control plane.
-    # In order to have these automatically trigger the correct pipelines they
-    # need to send their events with the correct log source.
-    if metadata[DD_SOURCE] == "eks":
-        if logs["logStream"].startswith("kube-apiserver-audit-"):
-            metadata[DD_SOURCE] = "kubernetes.audit"
-        elif logs["logStream"].startswith("kube-scheduler-"):
-            metadata[DD_SOURCE] = "kube_scheduler"
-        elif logs["logStream"].startswith("kube-apiserver-"):
-            metadata[DD_SOURCE] = "kube-apiserver"
-        elif logs["logStream"].startswith("kube-controller-manager-"):
-            metadata[DD_SOURCE] = "kube-controller-manager"
-        elif logs["logStream"].startswith("authenticator-"):
-            metadata[DD_SOURCE] = "aws-iam-authenticator"
-        # In case the conditions above don't match we maintain eks as the source
-
-    # Create and send structured logs to Datadog
-    for log in logs["logEvents"]:
-        yield merge_dicts(log, aws_attributes)
-
-
-def merge_dicts(a, b, path=None):
-    if path is None:
-        path = []
-    for key in b:
-        if key in a:
-            if isinstance(a[key], dict) and isinstance(b[key], dict):
-                merge_dicts(a[key], b[key], path + [str(key)])
-            elif a[key] == b[key]:
-                pass  # same leaf value
-            else:
-                raise Exception(
-                    "Conflict while merging metadatas and the log entry at %s"
-                    % ".".join(path + [str(key)])
-                )
-        else:
-            a[key] = b[key]
-    return a
 
 
 # Handle Cloudwatch Events
@@ -702,51 +179,3 @@ def normalize_events(events, metadata):
     )
 
     return normalized
-
-
-def get_state_machine_arn(message):
-    if message.get("execution_arn") is not None:
-        execution_arn = message["execution_arn"]
-        arn_tokens = re.split(r"[:/\\]", execution_arn)
-        arn_tokens[5] = "stateMachine"
-        return ":".join(arn_tokens[:7])
-    return ""
-
-
-# Lambda logs can be from either default or customized log group
-def process_lambda_logs(logs, aws_attributes, context, metadata):
-    lower_cased_lambda_function_name = get_lower_cased_lambda_function_name(logs)
-    if lower_cased_lambda_function_name is None:
-        return
-    # Split the arn of the forwarder to extract the prefix
-    arn_parts = context.invoked_function_arn.split("function:")
-    if len(arn_parts) > 0:
-        arn_prefix = arn_parts[0]
-        # Rebuild the arn with the lowercased function name
-        lower_cased_lambda__arn = (
-            arn_prefix + "function:" + lower_cased_lambda_function_name
-        )
-        # Add the lowe_rcased arn as a log attribute
-        arn_attributes = {"lambda": {"arn": lower_cased_lambda__arn}}
-        aws_attributes = merge_dicts(aws_attributes, arn_attributes)
-        env_tag_exists = (
-            metadata[DD_CUSTOM_TAGS].startswith("env:")
-            or ",env:" in metadata[DD_CUSTOM_TAGS]
-        )
-        # If there is no env specified, default to env:none
-        if not env_tag_exists:
-            metadata[DD_CUSTOM_TAGS] += ",env:none"
-
-
-# The lambda function name can be inferred from either a customized logstream name, or a loggroup name
-def get_lower_cased_lambda_function_name(logs):
-    logstream_name = logs["logStream"]
-    # function name parsed from logstream is preferred for handling some edge cases
-    function_name = get_lambda_function_name_from_logstream_name(logstream_name)
-    if function_name is None:
-        log_group_parts = logs["logGroup"].split("/lambda/")
-        if len(log_group_parts) > 1:
-            function_name = log_group_parts[1]
-        else:
-            return None
-    return function_name.lower()
