+++ aws/logs_monitoring/tests/test_lambda_function.py
@@ -31,7 +31,6 @@ from steps.enrichment import enrich
 from steps.transformation import transform
 from steps.splitting import split
 from steps.parsing import parse, parse_event_type
-
 env_patch.stop()
 
 
@@ -70,9 +69,10 @@ class TestInvokeAdditionalTargetLambdas(unittest.TestCase):
 
 
 class TestLambdaFunctionEndToEnd(unittest.TestCase):
-    @patch("enhanced_lambda_metrics.LambdaTagsCache.get_cache_from_s3")
-    def test_datadog_forwarder(self, mock_get_s3_cache):
-        mock_get_s3_cache.return_value = (
+    @patch("enhanced_lambda_metrics.LambdaTagsCache.get")
+    @patch("caching.cloudwatch_log_group_cache.CloudwatchLogGroupTagsCache.get_cache_from_s3")
+    def test_datadog_forwarder(self, mock_get_lambda_tags, mock_get_s3_cache):
+        mock_get_lambda_tags.return_value = (
             {
                 "arn:aws:lambda:sa-east-1:601427279990:function:inferred-spans-python-dev-initsender": [
                     "team:metrics",
@@ -84,6 +84,7 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
             },
             time(),
         )
+        mock_get_s3_cache.return_value = []
         context = Context()
         input_data = self._get_input_data()
         event = {
@@ -139,7 +140,8 @@ class TestLambdaFunctionEndToEnd(unittest.TestCase):
         del os.environ["DD_FETCH_LAMBDA_TAGS"]
 
     @patch("caching.cloudwatch_log_group_cache.CloudwatchLogGroupTagsCache.get")
-    def test_setting_service_tag_from_log_group_cache(self, cw_logs_tags_get):
+    @patch("caching.lambda_cache.LambdaTagsCache.get")
+    def test_setting_service_tag_from_log_group_cache(self, cw_logs_tags_get, lambda_tags_get):
         reload(sys.modules["settings"])
         reload(sys.modules["steps.parsing"])
         cw_logs_tags_get.return_value = ["service:log_group_service"]
