+++ aws/logs_monitoring/tests/test_parsing.py
@@ -1,43 +1,13 @@
-import base64
-import gzip
-import json
-from unittest.mock import MagicMock, patch
-import os
-import sys
 import unittest
-from approvaltests.approvals import verify_as_json
-from approvaltests.combination_approvals import verify_all_combinations
-from approvaltests.namer import NamerFactory
 
-sys.modules["trace_forwarder.connection"] = MagicMock()
-sys.modules["datadog_lambda.wrapper"] = MagicMock()
-sys.modules["datadog_lambda.metric"] = MagicMock()
-sys.modules["datadog"] = MagicMock()
-sys.modules["requests"] = MagicMock()
-sys.modules["requests_futures.sessions"] = MagicMock()
-
-env_patch = patch.dict(
-    os.environ,
-    {
-        "DD_API_KEY": "11111111111111111111111111111111",
-    },
-)
-env_patch.start()
-from steps.parsing import (
-    awslogs_handler,
+from steps.common import (
     parse_event_source,
-    parse_service_arn,
     get_service_from_tags_and_remove_duplicates,
-    get_state_machine_arn,
-    get_lower_cased_lambda_function_name,
-    get_structured_lines_for_s3_handler,
 )
 from settings import (
     DD_CUSTOM_TAGS,
     DD_SOURCE,
 )
-
-env_patch.stop()
 
 
 class TestParseEventSource(unittest.TestCase):
@@ -335,184 +305,6 @@ class TestParseEventSource(unittest.TestCase):
         self.assertEqual(parse_event_source({"Records": ["logs-from-s3"]}, ""), "s3")
 
 
-class TestParseServiceArn(unittest.TestCase):
-    def test_elb_s3_key_invalid(self):
-        self.assertEqual(
-            parse_service_arn(
-                "elb",
-                "123456789123/elasticloadbalancing/us-east-1/2022/02/08/123456789123_elasticloadbalancing_us-east-1_app.my-alb-name.123456789aabcdef_20220208T1127Z_10.0.0.2_1abcdef2.log.gz",
-                None,
-                None,
-            ),
-            None,
-        )
-
-    def test_elb_s3_key_no_prefix(self):
-        self.assertEqual(
-            parse_service_arn(
-                "elb",
-                "AWSLogs/123456789123/elasticloadbalancing/us-east-1/2022/02/08/123456789123_elasticloadbalancing_us-east-1_app.my-alb-name.123456789aabcdef_20220208T1127Z_10.0.0.2_1abcdef2.log.gz",
-                None,
-                None,
-            ),
-            "arn:aws:elasticloadbalancing:us-east-1:123456789123:loadbalancer/app/my-alb-name/123456789aabcdef",
-        )
-
-    def test_elb_s3_key_single_prefix(self):
-        self.assertEqual(
-            parse_service_arn(
-                "elb",
-                "elasticloadbalancing/AWSLogs/123456789123/elasticloadbalancing/us-east-1/2022/02/08/123456789123_elasticloadbalancing_us-east-1_app.my-alb-name.123456789aabcdef_20220208T1127Z_10.0.0.2_1abcdef2.log.gz",
-                None,
-                None,
-            ),
-            "arn:aws:elasticloadbalancing:us-east-1:123456789123:loadbalancer/app/my-alb-name/123456789aabcdef",
-        )
-
-    def test_elb_s3_key_multi_prefix(self):
-        self.assertEqual(
-            parse_service_arn(
-                "elb",
-                "elasticloadbalancing/my-alb-name/AWSLogs/123456789123/elasticloadbalancing/us-east-1/2022/02/08/123456789123_elasticloadbalancing_us-east-1_app.my-alb-name.123456789aabcdef_20220208T1127Z_10.0.0.2_1abcdef2.log.gz",
-                None,
-                None,
-            ),
-            "arn:aws:elasticloadbalancing:us-east-1:123456789123:loadbalancer/app/my-alb-name/123456789aabcdef",
-        )
-
-    def test_elb_s3_key_multi_prefix_gov(self):
-        self.assertEqual(
-            parse_service_arn(
-                "elb",
-                "elasticloadbalancing/my-alb-name/AWSLogs/123456789123/elasticloadbalancing/us-gov-east-1/2022/02/08"
-                "/123456789123_elasticloadbalancing_us-gov-east-1_app.my-alb-name.123456789aabcdef_20220208T1127Z_10"
-                ".0.0.2_1abcdef2.log.gz",
-                None,
-                None,
-            ),
-            "arn:aws-us-gov:elasticloadbalancing:us-gov-east-1:123456789123:loadbalancer/app/my-alb-name"
-            "/123456789aabcdef",
-        )
-
-
-class TestAWSLogsHandler(unittest.TestCase):
-    @patch("steps.parsing.CloudwatchLogGroupTagsCache.get")
-    @patch("steps.parsing.CloudwatchLogGroupTagsCache.release_s3_cache_lock")
-    @patch("steps.parsing.CloudwatchLogGroupTagsCache.acquire_s3_cache_lock")
-    @patch("steps.parsing.CloudwatchLogGroupTagsCache.write_cache_to_s3")
-    @patch("caching.base_tags_cache.send_forwarder_internal_metrics")
-    @patch("steps.parsing.CloudwatchLogGroupTagsCache.get_cache_from_s3")
-    def test_awslogs_handler_rds_postgresql(
-        self,
-        mock_get_s3_cache,
-        mock_forward_metrics,
-        mock_write_cache,
-        mock_acquire_lock,
-        mock_release_lock,
-        mock_cache_get,
-    ):
-        os.environ["DD_FETCH_LAMBDA_TAGS"] = "True"
-        os.environ["DD_FETCH_LOG_GROUP_TAGS"] = "True"
-        mock_acquire_lock.return_value = True
-        mock_cache_get.return_value = ["test_tag_key:test_tag_value"]
-        mock_get_s3_cache.return_value = (
-            {},
-            1000,
-        )
-
-        event = {
-            "awslogs": {
-                "data": base64.b64encode(
-                    gzip.compress(
-                        bytes(
-                            json.dumps(
-                                {
-                                    "owner": "123456789012",
-                                    "logGroup": "/aws/rds/instance/datadog/postgresql",
-                                    "logStream": "datadog.0",
-                                    "logEvents": [
-                                        {
-                                            "id": "31953106606966983378809025079804211143289615424298221568",
-                                            "timestamp": 1609556645000,
-                                            "message": "2021-01-02 03:04:05 UTC::@:[5306]:LOG:  database system is ready to accept connections",
-                                        }
-                                    ],
-                                }
-                            ),
-                            "utf-8",
-                        )
-                    )
-                )
-            }
-        }
-        context = None
-        metadata = {"ddsource": "postgresql", "ddtags": "env:dev"}
-
-        verify_as_json(list(awslogs_handler(event, context, metadata)))
-        verify_as_json(metadata, options=NamerFactory.with_parameters("metadata"))
-
-    @patch("steps.parsing.CloudwatchLogGroupTagsCache.get")
-    @patch("steps.parsing.StepFunctionsTagsCache.get")
-    @patch("steps.parsing.StepFunctionsTagsCache.release_s3_cache_lock")
-    @patch("steps.parsing.StepFunctionsTagsCache.acquire_s3_cache_lock")
-    @patch("steps.parsing.StepFunctionsTagsCache.write_cache_to_s3")
-    @patch("caching.base_tags_cache.send_forwarder_internal_metrics")
-    @patch("steps.parsing.StepFunctionsTagsCache.get_cache_from_s3")
-    def test_awslogs_handler_step_functions_tags_added_properly(
-        self,
-        mock_get_s3_cache,
-        mock_forward_metrics,
-        mock_write_cache,
-        mock_acquire_lock,
-        mock_release_lock,
-        mock_step_functions_cache_get,
-        mock_cw_log_group_cache_get,
-    ):
-        os.environ["DD_FETCH_LAMBDA_TAGS"] = "True"
-        os.environ["DD_FETCH_LOG_GROUP_TAGS"] = "True"
-        os.environ["DD_FETCH_STEP_FUNCTIONS_TAGS"] = "True"
-        mock_acquire_lock.return_value = True
-        mock_step_functions_cache_get.return_value = ["test_tag_key:test_tag_value"]
-        mock_cw_log_group_cache_get.return_value = []
-        mock_get_s3_cache.return_value = (
-            {},
-            1000,
-        )
-
-        event = {
-            "awslogs": {
-                "data": base64.b64encode(
-                    gzip.compress(
-                        bytes(
-                            json.dumps(
-                                {
-                                    "messageType": "DATA_MESSAGE",
-                                    "owner": "425362996713",
-                                    "logGroup": "/aws/vendedlogs/states/logs-to-traces-sequential-Logs",
-                                    "logStream": "states/logs-to-traces-sequential/2022-11-10-15-50/7851b2d9",
-                                    "subscriptionFilters": ["testFilter"],
-                                    "logEvents": [
-                                        {
-                                            "id": "37199773595581154154810589279545129148442535997644275712",
-                                            "timestamp": 1668095539607,
-                                            "message": '{"id":"1","type":"ExecutionStarted","details":{"input":"{"Comment": "Insert your JSON here"}","inputDetails":{"truncated":false},"roleArn":"arn:aws:iam::425362996713:role/service-role/StepFunctions-logs-to-traces-sequential-role-ccd69c03"},",previous_event_id":"0","event_timestamp":"1668095539607","execution_arn":"arn:aws:states:sa-east-1:425362996713:express:logs-to-traces-sequential:d0dbefd8-a0f6-b402-da4c-f4863def7456:7fa0cfbe-be28-4a20-9875-73c37f5dc39e"}',
-                                        }
-                                    ],
-                                }
-                            ),
-                            "utf-8",
-                        )
-                    )
-                )
-            }
-        }
-        context = None
-        metadata = {"ddsource": "postgresql", "ddtags": "env:dev"}
-
-        verify_as_json(list(awslogs_handler(event, context, metadata)))
-        verify_as_json(metadata, options=NamerFactory.with_parameters("metadata"))
-
-
 class TestGetServiceFromTags(unittest.TestCase):
     def test_get_service_from_tags(self):
         metadata = {
@@ -536,117 +328,6 @@ class TestGetServiceFromTags(unittest.TestCase):
         self.assertEqual(get_service_from_tags_and_remove_duplicates(metadata), "web")
         self.assertEqual(
             metadata[DD_CUSTOM_TAGS], "env:dev,tag,stack:aws:ecs,service:web,version:v1"
-        )
-
-
-class TestParsingStepFunctionLogs(unittest.TestCase):
-    def test_get_state_machine_arn(self):
-        invalid_sf_log_message = {"no_execution_arn": "xxxx/yyy"}
-        self.assertEqual(get_state_machine_arn(invalid_sf_log_message), "")
-
-        normal_sf_log_message = {
-            "execution_arn": "arn:aws:states:sa-east-1:425362996713:express:my-Various-States:7f653fda-c79a-430b-91e2-3f97eb87cabb:862e5d40-a457-4ca2-a3c1-78485bd94d3f"
-        }
-        self.assertEqual(
-            get_state_machine_arn(normal_sf_log_message),
-            "arn:aws:states:sa-east-1:425362996713:stateMachine:my-Various-States",
-        )
-
-        forward_slash_sf_log_message = {
-            "execution_arn": "arn:aws:states:sa-east-1:425362996713:express:my-Various-States/7f653fda-c79a-430b-91e2-3f97eb87cabb:862e5d40-a457-4ca2-a3c1-78485bd94d3f"
-        }
-        self.assertEqual(
-            get_state_machine_arn(forward_slash_sf_log_message),
-            "arn:aws:states:sa-east-1:425362996713:stateMachine:my-Various-States",
-        )
-
-        back_slash_sf_log_message = {
-            "execution_arn": "arn:aws:states:sa-east-1:425362996713:express:my-Various-States\\7f653fda-c79a-430b-91e2-3f97eb87cabb:862e5d40-a457-4ca2-a3c1-78485bd94d3f"
-        }
-        self.assertEqual(
-            get_state_machine_arn(back_slash_sf_log_message),
-            "arn:aws:states:sa-east-1:425362996713:stateMachine:my-Various-States",
-        )
-
-
-class TestLambdaCustomizedLogGroup(unittest.TestCase):
-    def test_get_lower_cased_lambda_function_name(self):
-        self.assertEqual(True, True)
-        # Non Lambda log
-        stepfunction_loggroup = {
-            "messageType": "DATA_MESSAGE",
-            "logGroup": "/aws/vendedlogs/states/logs-to-traces-sequential-Logs",
-            "logStream": "states/logs-to-traces-sequential/2022-11-10-15-50/7851b2d9",
-            "logEvents": [],
-        }
-        self.assertEqual(
-            get_lower_cased_lambda_function_name(stepfunction_loggroup), None
-        )
-        lambda_default_loggroup = {
-            "messageType": "DATA_MESSAGE",
-            "logGroup": "/aws/lambda/test-lambda-default-log-group",
-            "logStream": "2023/11/06/[$LATEST]b25b1f977b3e416faa45a00f427e7acb",
-            "logEvents": [],
-        }
-        self.assertEqual(
-            get_lower_cased_lambda_function_name(lambda_default_loggroup),
-            "test-lambda-default-log-group",
-        )
-        lambda_customized_loggroup = {
-            "messageType": "DATA_MESSAGE",
-            "logGroup": "customizeLambdaGrop",
-            "logStream": "2023/11/06/test-customized-log-group1[$LATEST]13e304cba4b9446eb7ef082a00038990",
-            "logEvents": [],
-        }
-        self.assertEqual(
-            get_lower_cased_lambda_function_name(lambda_customized_loggroup),
-            "test-customized-log-group1",
-        )
-
-
-class TestS3EventsHandler(unittest.TestCase):
-    def parse_lines(self, data, key, source):
-        bucket = "my-bucket"
-        gzip_data = gzip.compress(bytes(data, "utf-8"))
-
-        return [
-            l
-            for l in get_structured_lines_for_s3_handler(gzip_data, bucket, key, source)
-        ]
-
-    def test_get_structured_lines_waf(self):
-        key = "mykey"
-        source = "waf"
-        verify_all_combinations(
-            lambda d: self.parse_lines(d, key, source),
-            [
-                [
-                    '{"timestamp": 12345, "key1": "value1", "key2":"value2"}\n',
-                    '{"timestamp": 12345, "key1": "value1", "key2":"value2"}\n{"timestamp": 789760, "key1": "value1", "key3":"value4"}\n',
-                    '{"timestamp": 12345, "key1": "value1", "key2":"value2" "key3": {"key5" : "value5"}}\r{"timestamp": 12345, "key1": "value1"}\n',
-                    '{"timestamp": 12345, "key1": "value1", "key2":"value2" "key3": {"key5" : "value5"}}\f{"timestamp": 12345, "key1": "value1"}\n',
-                    '{"timestamp": 12345, "key1": "value1", "key2":"value2" "key3": {"key5" : "value5"}}\u00A0{"timestamp": 12345, "key1": "value1"}\n',
-                    "",
-                    "\n",
-                ]
-            ],
-        )
-
-    def test_get_structured_lines_cloudtrail(self):
-        key = (
-            "123456779121_CloudTrail_eu-west-3_20180707T1735Z_abcdefghi0MCRL2O.json.gz"
-        )
-        source = "cloudtrail"
-        verify_all_combinations(
-            lambda d: self.parse_lines(d, key, source),
-            [
-                [
-                    '{"Records": [{"event_key" : "logs-from-s3"}]}',
-                    '{"Records": [{"event_key" : "logs-from-s3"}, {"key1" : "data1", "key2" : "data2"}]}',
-                    '{"Records": {}}',
-                    "",
-                ]
-            ],
         )
